Improve type safeness of InitializedReport
