+++ crates/daphne/src/lib.rs
@@ -91,7 +91,7 @@ use url::Url;
 use vdaf::mastic::MasticWeight;
 
 pub use messages::request::{DapRequest, DapRequestMeta, DapResponse};
-pub use protocol::aggregator::InitializedReport;
+pub use protocol::report_init::{InitializedReport, WithPeerPrepShare};
 
 /// DAP version used for a task.
 #[derive(Clone, Copy, Debug, Default, Deserialize, Eq, Hash, PartialEq, Serialize)]
